ml-services/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
moviepy>=1.0.3
cloudinary>=1.33.0
gtts>=2.3.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
boto3>=1.26.0
celery>=5.3.0
redis>=4.6.0
python-magic>=0.4.27
validators>=0.20.0
Werkzeug>=2.3.0
click>=8.0.0

# Text-to-Speech services
gTTS>=2.3.0
pyttsx3>=2.90
azure-cognitiveservices-speech>=1.30.0

# AI/ML packages (optional, for advanced features)
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
speechbrain>=0.5.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
scikit-video>=1.1.11

# Background processing
dramatiq>=1.15.0
dramatiq[redis]>=1.15.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0